+++ src/main/java/com/mustache/bbs/service/HospitalService.java
@@ -15,7 +15,7 @@ public class HospitalService {
         this.hospitalRepository = hospitalRepository;
     }
 
-    public HospitalResponse getHospital(Integer id) {
+    public HospitalResponse getHospital(Long id) {
         Optional<Hospital> optHospital = hospitalRepository.findById(id); // Entity
         Hospital hospital = optHospital.get();
         HospitalResponse hospitalResponse = Hospital.of(hospital); //Dto
